+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/BlobListDetails.java
@@ -23,6 +23,8 @@ public final class BlobListDetails {
     private boolean retrieveUncommittedBlobs;
     private boolean retrieveDeletedBlobs;
     private boolean retrieveVersions;
+    private boolean retrieveImmutabilityPolicy;
+    private boolean retrieveLegalHold;
 
     /**
      * Constructs an unpopulated {@link BlobListDetails}.
@@ -173,6 +175,46 @@ public final class BlobListDetails {
     }
 
     /**
+     * Whether immutability policy for the blob should be returned.
+     *
+     * @return a flag indicating if immutability policy for the blob will be returned in the listing
+     */
+    public boolean getRetrieveImmutabilityPolicy() {
+        return retrieveImmutabilityPolicy;
+    }
+
+    /**
+     * Whether immutability policy for the blob should be returned.
+     *
+     * @param retrieveImmutabilityPolicy Flag indicating whether immutability policy for the blob should be returned
+     * @return the updated BlobListDetails object
+     */
+    public BlobListDetails setRetrieveImmutabilityPolicy(boolean retrieveImmutabilityPolicy) {
+        this.retrieveImmutabilityPolicy = retrieveImmutabilityPolicy;
+        return this;
+    }
+
+    /**
+     * Whether legal hold for the blob should be returned.
+     *
+     * @return a flag indicating if legal hold for the blob will be returned in the listing
+     */
+    public boolean getRetrieveLegalHold() {
+        return retrieveLegalHold;
+    }
+
+    /**
+     * Whether legal hold for the blob should be returned.
+     *
+     * @param retrieveLegalHold Flag indicating whetherlegal hold for the blob  should be returned
+     * @return the updated BlobListDetails object
+     */
+    public BlobListDetails setRetrieveLegalHold(boolean retrieveLegalHold) {
+        this.retrieveLegalHold = retrieveLegalHold;
+        return this;
+    }
+
+    /**
      * @return a list of the flag set to true
      */
     public ArrayList<ListBlobsIncludeItem> toList() {
@@ -197,6 +239,12 @@ public final class BlobListDetails {
         }
         if (this.retrieveVersions) {
             details.add(ListBlobsIncludeItem.VERSIONS);
+        }
+        if (this.retrieveImmutabilityPolicy) {
+            details.add(ListBlobsIncludeItem.IMMUTABILITYPOLICY);
+        }
+        if (this.retrieveLegalHold) {
+            details.add(ListBlobsIncludeItem.LEGALHOLD);
         }
         return details;
     }