+++ sdk/storage/azure-storage-blob/src/test/java/com/azure/storage/blob/ImmutableStorageWithVersioningTest.groovy
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.storage.blob
+
+import com.azure.storage.common.test.shared.extensions.RequiredServiceVersion
+
+class ImmutableStorageWithVersioningTest extends APISpec {
+
+
+}